S_KaladaiYT/Level2_QuizPrep/Session_09_Personal_CursorAI/pyproject.toml
[project]
name = "session-09-personal-cursorai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents==0.2.11",
    "pydantic>=2.12.2",
    "python-dotenv>=1.1.1",
    "rich>=14.2.0",
]
